feat(oauth2) add ability to persist refresh tokens
Adds a configuration option that will persist the
refresh token value throughout the lifecycle of
the token. If `persistent_refresh_token` is
true, the refresh token generated with the first
access token will be used forever. RFC6749 states
that the refresh token MAY be regenerated, but
it is not required. In the case of a refresh, if
a client fails to receive the new access token
and a newly generated refresh token (maybe due to
a timeout) the client is left without the ability
to recover and must re-authenticate. When using a
persistent refresh token, the refresh can be
re-attempted.

https://tools.ietf.org/html/rfc6749#section-6
https://discuss.konghq.com/t/long-lived-refresh-tokens/659
https://discuss.konghq.com/t/long-lived-oauth-2-refresh-tokens/3005